refactor: WI-0300 payroll service modular split phase3
