+++ ROADMAP.md
@@ -570,3 +570,4 @@ Phase 8: Extensions (ATS, performance, expenses, analytics)
 - WI-0297 payroll decomposition phase2 baseline(`src/features/payroll/year-end-audit-payload-helpers.ts` extraction for settlement-hash and filing/finalization audit payload parsers + `src/features/payroll/service.ts` wrapper rewiring + WI-0297 regression test)
 - WI-0298 admin decomposition phase1 baseline(src/app/admin/page-queue-helpers.ts extraction for approval-queue wait/SLA/filter/search-sort derived logic + src/app/admin/page.tsx rewiring with no behavior change + WI-0298 regression test)
 - WI-0299 admin locale dynamic UI gap fix baseline(/admin locale-aware queue/status/log/fallback text wiring via useI18n + src/app/admin/page-queue-helpers.ts queueLabel locale parameterization + WI-0299 regression test)
+- WI-0300 payroll decomposition phase3 baseline(src/features/payroll/kr-statutory-helpers.ts extraction for KR statutory tax/insurance normalize/calc helpers + src/features/payroll/service.ts wrapper rewiring + WI-0300 regression test)
